--- a/src/main/scala/mesosphere/marathon/core/appinfo/PodStatusService.scala
+++ b/src/main/scala/mesosphere/marathon/core/appinfo/PodStatusService.scala
@@ -12,4 +12,9 @@ trait PodStatusService {
     * @return the status of the pod at the given path, if such a pod exists
     */
   def selectPodStatus(id: PathId, selector: PodSelector = Selector.all): Future[Option[PodStatus]]
+
+  /**
+    * @return the statuses of the pods at the given paths, if the pod exists
+    */
+  def selectPodStatuses(ids: Set[PathId], selector: PodSelector = Selector.all): Future[Seq[PodStatus]]
 }